+++ sdk/monitor/azure-opentelemetry-exporter-azuremonitor/azure/opentelemetry/exporter/azuremonitor/__init__.py
@@ -4,9 +4,10 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------
 
+from azure.opentelemetry.exporter.azuremonitor.export.metrics._exporter import AzureMonitorMetricsExporter
 from azure.opentelemetry.exporter.azuremonitor.export.trace._exporter import AzureMonitorTraceExporter
 from azure.opentelemetry.exporter.azuremonitor._options import ExporterOptions
 from ._version import VERSION
 
-__all__ = ["AzureMonitorTraceExporter", "ExporterOptions"]
+__all__ = ["AzureMonitorMetricsExporter", "AzureMonitorTraceExporter", "ExporterOptions"]
 __version__ = VERSION